Rename "vipsdispapp.{c,h}" as "app.{c,h}"
In line with commit libvips/nip4@3aeeac7.
