+++ src/app.h
@@ -1,5 +1,5 @@
-#ifndef __VIPSDISP_APP_H
-#define __VIPSDISP_APP_H
+#ifndef __APP_H
+#define __APP_H
 
 #define APP_TYPE (vipsdisp_app_get_type())
 G_DECLARE_FINAL_TYPE(VipsdispApp, vipsdisp_app, VIPSDISP, APP, GtkApplication)
@@ -8,4 +8,4 @@ G_DECLARE_FINAL_TYPE(VipsdispApp, vipsdisp_app, VIPSDISP, APP, GtkApplication)
 
 VipsdispApp *vipsdisp_app_new(void);
 
-#endif /* __VIPSDISP_APP_H */
+#endif /*__APP_H*/
